refactored route, controller, service
