+++ app.js
@@ -6,8 +6,9 @@ const logger = require("morgan");
 const cors = require("cors");
 
 const indexRouter = require("./routes/index");
-// const usersRouter = require('./routes/users');
-const apiRouter = require("./routes/api");
+const usersRouter = require("./routes/users");
+const hobbyRouter = require("./routes/hobby");
+const commentRouter = require("./routes/comment");
 
 const app = express();
 
@@ -25,7 +26,9 @@ app.use(express.static(path.join(__dirname, "public")));
 
 app.use("/", indexRouter);
 // app.use('/users', usersRouter);
-app.use("/api", apiRouter);
+app.use("/api/hobby", hobbyRouter);
+app.use("/api/comment", commentRouter);
+app.use("/api/user", usersRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
