+++ services/hobby.js
@@ -84,54 +84,54 @@ exports.postComment = async function (
   return comment;
 };
 
-exports.putComment = async function (reqCommentId, reqContent) {
-  // 最初に fineOne で 行を特定する
-  const comment = await Comment.findOne({
-    where: {
-      id: reqCommentId,
-    },
-  });
+// exports.putComment = async function (reqCommentId, reqContent) {
+//   // 最初に fineOne で 行を特定する
+//   const comment = await Comment.findOne({
+//     where: {
+//       id: reqCommentId,
+//     },
+//   });
 
-  // 特定した行に対して、update を実行する。
-  const updated = await comment.update({
-    content: reqContent,
-  });
+//   // 特定した行に対して、update を実行する。
+//   const updated = await comment.update({
+//     content: reqContent,
+//   });
 
-  return updated;
-};
+//   return updated;
+// };
 
-exports.deleteComment = async function (reqCommentId) {
-  const comment = await Comment.findOne({
-    where: {
-      id: reqCommentId,
-    },
-  });
-  try {
-    // const result = await comment.destroy();
-    await comment.destroy();
+// exports.deleteComment = async function (reqCommentId) {
+//   const comment = await Comment.findOne({
+//     where: {
+//       id: reqCommentId,
+//     },
+//   });
+//   try {
+//     // const result = await comment.destroy();
+//     await comment.destroy();
 
-    return;
-  } catch (err) {
-    console.log(err);
-  }
-};
+//     return;
+//   } catch (err) {
+//     console.log(err);
+//   }
+// };
 
-exports.getUser = async function (reqName, reqPassword) {
-  const result = await NacomerUser.findOne({
-    raw: true,
-    where: {
-      name: reqName,
-      password: reqPassword,
-    },
-    attributes: ["id", "name"],
-  });
-  return result;
-};
+// exports.getUser = async function (reqName, reqPassword) {
+//   const result = await NacomerUser.findOne({
+//     raw: true,
+//     where: {
+//       name: reqName,
+//       password: reqPassword,
+//     },
+//     attributes: ["id", "name"],
+//   });
+//   return result;
+// };
 
-exports.postUser = async function (req) {
-  const user = await NacomerUser.create({
-    name: req.name,
-    password: req.password,
-  });
-  return user;
-};
+// exports.postUser = async function (req) {
+//   const user = await NacomerUser.create({
+//     name: req.name,
+//     password: req.password,
+//   });
+//   return user;
+// };
